Add files via upload
I ran the Python script ceasar.py, which tries all possible shifts from 1 to 25 to decode the Caesar cipher. The results for each shift were printed, and at shift 14, I got the correct decoded message:

"This cipher is not secure, we will need to improve it. HSCTF{c43sar-c1ph3r}".

So, shift 14 was the right one to crack the cipher, and I found the flag: HSCTF{c43sar-c1ph3r}.